init: add new dependencies. Init wss
